+++ src/types.ts
@@ -52,6 +52,7 @@ export interface Profile {
   type: ProfileTypes;
   uuid: string;
 }
+
 export interface Email {
   recipients: string;
   subject: string;
